Add new `applications` module
...and nuke `testIntegration`.
