+++ analyzer/src/test/resources/com/github/cubuspl42/sigmaLang/analyzer/semantics/imports/main.sigma
