+++ applications/src/main/kotlin/com/github/cubuspl42/sigmaLang/applications/utils.kt
@@ -1,0 +1,7 @@
+package com.github.cubuspl42.sigmaLang.applications
+
+import com.github.cubuspl42.sigmaLang.analyzer.evaluation.values.*
+
+private fun getResourceAsText(
+    path: String,
+): String? = object {}.javaClass.getResource(path)?.readText()
