+++ applications/src/main/resources/com/github/cubuspl42/sigmaLang/applications/euler/problem10.sigma
