+++ applications/src/main/resources/com/github/cubuspl42/sigmaLang/applications/euler/problem14.sigma
