+++ applications/src/main/resources/com/github/cubuspl42/sigmaLang/applications/html/main.sigma
