+++ gradle/libs.versions.toml
@@ -1,6 +1,7 @@
 [versions]
 # libraries
 annotations = "24.0.1"
+kotlinxSerialization = "1.6.0"
 
 # plugins
 kotlin = "1.9.10"
@@ -11,6 +12,7 @@ kover = "0.7.3"
 
 [libraries]
 annotations = { group = "org.jetbrains", name = "annotations", version.ref = "annotations" }
+kotlinxSerialization = { group = "org.jetbrains.kotlinx", name = "kotlinx-serialization-json", version.ref = "kotlinxSerialization" }
 
 [plugins]
 changelog = { id = "org.jetbrains.changelog", version.ref = "changelog" }
