+++ settings.gradle.kts
@@ -1,4 +1,5 @@
 rootProject.name = "sigma-lang"
 
 include("analyzer")
+include("applications")
 include("intellij-plugin")
